Moving administration logic to the AdminDashboard module
Adding new controllers and views to the AdminDashboard module with the existing logic from the Administration area of Restaurant.Web.

Refactoring the OrderNumberValidation attribute and its tests to account for the change in administration logic.

Adding a new configuration - WebRootPath - the path to the wwwroot folder of Restaurant.Web project, to the appsettings.json file of the AdminDashboard as a temporary solution to adding meal images through the new module.